+++ src/main/java/de/slikey/effectlib/util/ParticleEffect.java
@@ -1409,4 +1409,19 @@ public enum ParticleEffect {
             display(offsetX, offsetY, offsetZ, speed, amount, center, range);
         }
     }
+
+    public ParticleData getData(Material material, Byte blockData) {
+        ParticleData data = null;
+        if (this == ParticleEffect.BLOCK_CRACK || this == ParticleEffect.ITEM_CRACK || this == ParticleEffect.BLOCK_DUST) {
+            if (material != null && material != Material.AIR) {
+                if (this == ParticleEffect.ITEM_CRACK) {
+                    data = new ParticleEffect.ItemData(material, blockData);
+                } else {
+                    data = new ParticleEffect.BlockData(material, blockData);
+                }
+            }
+        }
+
+        return data;
+    }
 }